backend/pyproject.toml
[project]
name = "scensei-backend"
version = "0.1.0"
description = "Scensei - 香水ソムリエAIエージェント"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore>=1.2.0",
    "bedrock-agentcore-starter-toolkit>=0.2.8",
    "strands-agents>=1.23.0",
    "strands-agents-tools>=0.2.19",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
